==> 00_sp_complier.Rmd <==
---
title: "00_sp_complier"
author: "Caroline McKeon"
date: "03/17/2020"
output: word_document
---

This markdown is for the extraction, wrangling and analysis of data relating to species spatial pattern (range metrics) and climate and human footprint.

# Set up
```{r setup, include=FALSE}
#install.packages(c("tidyverse", "glmmTMB"))
#install.packages("RPostgreSQL", type="source") 
#install.packages("BIEN") 
source("00_sp_functions.R")
knitr::opts_chunk$set(include=FALSE) 
knitr::opts_chunk$set(fig.width=10, fig.height=7) 

## Quick look at model dataframe
## Factors
# for (i in names(Filter(is.factor, ModelDF))) {
#   plot(ModelDF[,i], 
#        main = paste(i))
# }
# 
# ## Numeric variables
# for (i in names(Filter(is.numeric, ModelDF))) {
#   hist(ModelDF[,i], 
#        breaks = 3000, 
#        main = paste(i), 
#        xlab = paste(i))
# }
```

## get range metrics
Read in and clean species range metric data
```{r}
metrics <- read.csv("GRSmetrics_AFE_pol_50km_pol_line_mcp.csv") ## metrics provided by Anna Csergo in Spring 2022. 4260 obs of 42 vars
metrics <- metrics[metrics$Model == "Occurrence",] ## 810 species
metrics <- metrics[which(!is.na(metrics$Species)),]  ## 809 species
metrics$Species <- gsub(" ssp.*", "", metrics$Species) ## remove subspecies (for now). 714 unique species
metrics$Species <- gsub("[[:space:]]*$", "", metrics$Species) ##  692 unique species
small <- c(" s.lat.", " s.str.")
metrics$Species <- gsub(paste(small, collapse="|"), "", metrics$Species) ## 691
metrics$Species <- gsub("//.", "", metrics$Species) 

# Quercus petraea.
# Ranunculus montanus.
# Pinus uncinata.

metrics <- unique(metrics) # 805 obs of 42 vars
metrics <- unique(metrics[, which(names(metrics) %in% c("Species", "total.area", "range.size", "effective.mesh.size", 
                                                        "mean.shape.index", "prop.landscape", "perimeter.area.frac.dim"))])
metrics <- metrics[c("Species", "total.area", "range.size", "effective.mesh.size", "prop.landscape","mean.shape.index", "perimeter.area.frac.dim")]
metrics <- unique(metrics) # 810 obs of 7 vars
names(metrics) <- c("species","Occupied area", "Geographic range size", "Patch size distribution", "Geographic range filling", "Patch shape complexity", "Geographic range fractality") 

## tidy names
metrics$species[grep("oslo", metrics$species, useBytes = TRUE)] <- "Saxifraga osloensis"
metrics$species[metrics$species == "Salix repens coll."] <- "Salix repens"

metrics$species <- factor(metrics$species)
levels(metrics$species) <- gsub(" ", "_", levels(metrics$species))
```

```{r}
## Numeric variables
for (i in names(Filter(is.numeric, metrics))) {
  hist(metrics[,i],
       breaks = 100,
       main = paste(i),
       xlab = paste(i))
}
```

## get phylogeny
```{r}
if(!exists("clean_tree")) {
  try(clean_tree <- read.tree("Data_sp_clean_tree.tre"))
  } else source("01_clean_phylogeny.R")

## 653 species with metrics and phylogeny April 2022
```

## 
```{r}
## load environmental data ----------------
# mat <- raster("wc2/wc2.1_30s_bio_1.tif") ## mean annual temperature (C*10)
# map <- raster("wc2/wc2.1_30s_bio_12.tif") ## mean annual precipatation (mm)
# map_var <- raster("wc2/wc2.1_30s_bio_15.tif")  ## mean annual precip coeff variation
# mat_var <- raster("wc2/wc2.1_30s_bio_4.tif") ## mean annual temp SD*100
# 
# ## make climate variables into one object (raster brick)
# clim_map <- brick(map, mat, map_var, mat_var) 

## get occurrence data
# endem <- read.delim2("AFE_endemics.txt")
# endem <- endem[, which(names(endem) %in% c("afe", "lon", "lat"))]
# endem$lon <- as.numeric(as.character(endem$lon))
# endem$lat <- as.numeric(as.character(endem$lat))
# endem$afe <- gsub(" ssp.*", "", endem$afe)
# endem$afe <- gsub(" ssp.*", "", endem$afe)
# endem$afe <- gsub(" s.lat.", "", endem$afe)
# endem$afe <- gsub(" s.str.", "", endem$afe)
# endem$afe <- gsub("coll.", "coll", endem$afe)
# endem$afe[endem$afe == "Saxifraga osloënsis"] <- "Saxifraga osloensis"
# endem$afe <- as.character(endem$afe)
# endem$afe[endem$afe == "Arabis collna"] <- "Arabis collina"
# endem$afe[endem$afe == "Dianthus collnus"] <- "Dianthus collinus"
# names(endem) <- c("species", "y", "x") 
# sp <- endem
# 
# sp <- sp[, c("species", "x", "y")] ## 735 unique species
# 
# ## make dataframe with just the lat and long co-ordinates of data that is relevant to my analysis
# sp_co <- sp %>% .[, which(names(.) %in% c("x", "y"))]
# 
# ## #extract climate values for coordinates in (full) dataset
# full_clim <- data.frame(raster::extract(clim_map,sp_co)) 
# ## create dataset with both climate values and co-ordinates of the values
# full_clim <- cbind(full_clim,sp)
# env <- unique(full_clim)
# names(env) <- c("map", "mat", "map_var", "mat_var", "species", "Longitude", "Latitude") 
# env <- drop_na(env)
#saveRDS(env, "Data_occurences_climate_values.rds")
```


## get current climate, humanfootprint and climate change velocity means across all species total areas
```{r}
## once you have run through script 02_remote_a once interactively, you can run all of these scripts in the terminal 

# setwd("/Users/macbookpro/Library/CloudStorage/OneDrive-Personal/PhD/spatialpattern_climate_humanfootprint")

## load base rasters and occurrence data
#this script will need to be interactive the first time you run it to create all the objects that can be easily sourced later
#source("02_remote_a_load_hf_vel_clim.R")

## for all species make 1km2 resolution occurrence rasters based on the AFE irregular grids cells 
#source("02_remote_b_make_rasters.R")
#source("02_remote_c_make_rasters.R")

## get mean values for hf, vel and climate across each species' occurrence
#source("02_remote_d_get_ratios.R")

```

have a quick look at everything

```{r}
rat <- readRDS("Data_ratios_dataframe.rds")

for(i in names(Filter(is.numeric, rat))) {
  hist(rat[,i],
       breaks = 50,
       main = paste(i),
       xlab = paste(i))
}
```


## these scripts is also interactive, so work your way through them  

#run analyses
# 03_sp_analysis.R

#plot results
# 04_sp_results.Rmd
# 05_maps.Rmd
