==> 00_str_tr8_complier.Rmd <==
---
title: "00_str_tr8_complier"
author: "Caroline McKeon"
date: "03/17/2020"
output: word_document
---

This markdown is for the extraction, wrangling and analysis of data relating to species range metrics and functional traits.

# Notes: 
- Anna Csergo is already looking at the relationship between spatial scales in these metrics across species. 
- How rare and how recorded a species is will also matter.
- which traits to use?
- species mean or location specific traits?
- Measures of intraspecific trait variability?
- could re-extract and clean up massively (amalgamate trait data), might improve sample size

# Set up
```{r setup, include=FALSE}
#install.packages(c("tidyverse", "glmmTMB"))
#install.packages("RPostgreSQL", type="source")
#install.packages("BIEN")
source("00_str_functions.R")
knitr::opts_chunk$set(include=FALSE) 
knitr::opts_chunk$set(fig.width=16, fig.height=10) 
```

## get traits
Get TRY and BIEN trait data
```{r}
source("01_get_trait_data.R")
```
## clean traits
Clean and collate trait data
```{r}
source("02_clean_trait_data.R")
```
## get Climate values (niche breadth)
```{r}
# source("03a_str_nichebreadth_pca.R")
nb <- readRDS("Data_nichebreadth.rds")
```

## get range metrics
Read in and clean species range metric data
```{r}
metrics <- read.csv("Data_range_metrics.csv") ## metrics provided by Anna Csergo in spring 2019
metrics$Species <- gsub(" ssp.*", "", metrics$Species) ## 701 unique species
metrics <- metrics[metrics$Model == "Occurrence",]
metrics <- unique(metrics[, which(names(metrics) %in% c("Species", "total.area", "range.size", "effective.mesh.size", 
                                                        "mean.shape.index", "prop.landscape", "perimeter.area.frac.dim"))])
metrics <- metrics[c("Species", "total.area", "range.size", "effective.mesh.size", "mean.shape.index", "prop.landscape", "perimeter.area.frac.dim")]
names(metrics) <- c("species", "total.area", "range.size", "effective.mesh.size", "mean.shape.index", "prop.landscape", "perimeter.area.frac.dim")

metrics$species[grep("oslo", metrics$species, useBytes = TRUE)] <- "Saxifraga osloensis"
metrics <- unique(merge(metrics, nb[,which(names(nb) %in% c('sp.list', "nb"))], by.x = "species", by.y = "sp.list"))
metrics <- metrics[metrics$species != "Helleborus lividus",] ## had infinity value

## log and scale raw data
metrics$perimeter.area.frac.dim <- (metrics$perimeter.area.frac.dim + sqrt(min(metrics$perimeter.area.frac.dim, na.rm = T)^2)) + 1
for (i in names(Filter(is.numeric, metrics[, which(names(metrics) %nin% c("mean.shape.index", "prop.landscape", "nb"))]))) {
        metrics[, i] <- c(log(metrics[,i]))
        }
for (i in names(Filter(is.numeric, metrics))) {
        metrics[, i] <- c(scale(metrics[,i]))
        }
length(unique(metrics$species)) ## 700 unique species
```

## get phylogeny
Set up phylogeny
```{r}
if(!exists("clean_tree")) {
  try(clean_tree <- read.tree("Data_str_clean_tree.tre"))
  } else source("03b_clean_phylogeny.R")
```



## get trait sumamry statistics
Calculate summary stats for continuous traits
```{r}
cont <- c("height", "leaf_area", "seed_mass", "sla")
## get the single values for each species
for(j in cont){
  a <- x[[j]]
  a[,2] <- as.numeric(a[,2])
  
  for(i in levels(a$species)){
  a$mean[a$species == i] <- gm_mean(a[,2][a$species == i])
  a$median[a$species == i] <- median(a[,2][a$species == i])
  a$max[a$species == i] <- max(a[,2][a$species == i])
  a$coeff_var[a$species == i] <- (sd(a[,2][a$species == i])/mean(a[,2][a$species == i]))*100
  }
  
  x[[j]] <- unique(a[,-2])
}

rm(a)

## Extract each dataset
height <- x[["height"]] ## 195 species
leaf_area <- x[["leaf_area"]] ## 41 species
lifeform <- x[["lifeform"]] ## 165 species (40610?!)
seed_mass <- x[["seed_mass"]] ## 96 species
sla <- x[["sla"]] ## 50 species
woodiness <- x[["woodiness"]] ## 238 species

lifeform$lifeform <- factor(lifeform$lifeform, levels = c("phanerophyte", "chamaephyte", "hemicryptophyte", "cryptophyte", "therophyte"))
```


Look at overlap in species between trait datasets
```{r eval=FALSE, include=FALSE}
# for maybe if you wanted this list all together in a data frame
b <- data.frame(unique(mydata$species))
names(b) <- "species"
# for (i in levels(mydata$trait_name)){
#   b <- unique(merge(b, mydata[,which(names(mydata) %in% c('species', i))], by = "species"))
# }

b <- droplevels(unique(merge(woodiness, lifeform, by = "species"))) ## 140
b <- droplevels(unique(merge(b, height, by = "species"))) ## 137
b <- droplevels(unique(merge(b, seed_mass, by = "species"))) ## 80
b <- droplevels(unique(merge(b, sla, by = "species"))) ## 30
b <- droplevels(unique(merge(b, leaf_area, by = "species"))) ## 30
b <- drop_na(b)
rm(b)
```


Plot summary stats
```{r}
stat <- c("mean","median", "max", "coeff_var")
par(mfrow = c(4,4))
  for(j in stat){
for(i in cont){

    hist(log(x[[i]][,j]), main = paste(i,j), xlab = paste(j), cex = 1, breaks = 50)
}}
```


```{r}
par(mfrow = c(1,2))
plot(factor(lifeform$lifeform))
plot(factor(woodiness$woodiness))
```

Merge trait and metric dataframes

```{r}
b <- data.frame(unique(mydata$species))
names(b) <- "species"
for (i in cont){
  b <- merge(b, x[[i]][,which(names(x[[i]]) %in% c('species', "max"))], by = "species", all.x = T); 
  names(b)[names(b) == 'max'] <- paste(i, "max", sep = "_") 
}
for (i in c("lifeform", "woodiness")){
  b <- merge(b, x[[i]], by = "species", all.x = T)
}

mydata <- unique(merge(metrics, b, by = "species"))
mydata$woodiness <- factor(mydata$woodiness)
mydata$lifeform <- factor(mydata$lifeform)
# ## get ModelDF species names in the right format to match with those in phylogenetic trees
mydata$species <- factor(mydata$species)
levels(mydata$species) <- gsub(" ", "_", levels(mydata$species))
mydata$lifeform <- factor(mydata$lifeform, levels = c("phanerophyte", "chamaephyte", "hemicryptophyte", "cryptophyte", "therophyte"))
rm(b)


```


```{r}
## Quick look at model dataframe
## Factors
for (i in names(Filter(is.factor, mydata))) {
  plot(mydata[,i],
       main = paste(i))
}

## Numeric variables
for (i in names(Filter(is.numeric, mydata))) {
  hist((mydata[,i]),
       breaks = 3000,
       main = paste(i),
       xlab = paste(i))
}
```


Set up individual datasets for analysis
```{r}
## Extract each dataset
height <- unique(drop_na(droplevels(mydata[, which(names(mydata) %in% c("species", "total.area", "range.size", "effective.mesh.size", "mean.shape.index", 
                                                      "prop.landscape", "perimeter.area.frac.dim", "nb", "height_max"))])))
leaf_area <- unique(drop_na(droplevels(mydata[, which(names(mydata) %in% c("species", "total.area", "range.size", "effective.mesh.size", "mean.shape.index", 
                                                      "prop.landscape", "perimeter.area.frac.dim","nb", "leaf_area_max"))])))
lifeform <- unique(drop_na(droplevels(mydata[, which(names(mydata) %in% c("species","total.area", "range.size", "effective.mesh.size", "mean.shape.index", 
                                                      "prop.landscape", "perimeter.area.frac.dim", "nb", "lifeform"))])))
seed_mass <- unique(drop_na(droplevels(mydata[, which(names(mydata) %in% c("species", "total.area", "range.size", "effective.mesh.size", "mean.shape.index", 
                                                      "prop.landscape", "perimeter.area.frac.dim", "nb", "seed_mass_max"))])))
sla <- unique(drop_na(droplevels(mydata[, which(names(mydata) %in% c("species", "total.area", "range.size", "effective.mesh.size", "mean.shape.index", 
                                                      "prop.landscape", "perimeter.area.frac.dim", "nb", "sla_max"))])))
woodiness <- unique(drop_na(droplevels(mydata[, which(names(mydata) %in% c("species", "total.area", "range.size", "effective.mesh.size", "mean.shape.index", 
                                                      "prop.landscape", "perimeter.area.frac.dim", "nb", "woodiness"))])))
## log transform and standardize raw trait data
height$height_max <- c(scale(log(height$height_max)))
leaf_area$leaf_area_max <- c(scale(log(leaf_area$leaf_area_max)))
seed_mass$seed_mass_max <- c(scale(log(seed_mass$seed_mass_max)))
sla$sla_max <- c(scale(log(sla$sla_max)))

```



## look at colinearity
## trait ~ trait
## metric ~ metric
```{r, fig.width=10, fig.height=9}
source("04_exploration_and_colinearity.R")
```


## Analysis
## 1. Metric ~ trait (*climate)

```{r}
## Can source these scripts, but since you have the model object lists and the effect plots saved, generally easier to look at these as needed.
# source("05a_str_height_analysis.R")
# source("05b_str_bayesian_bivariate_analysis.R")
# source("05c_str_bayesian_bivariate_analysis.R")
# source("05d_str_sla_analysis.R")
# source("05e_str_lifeform_analysis.R")
# source("05f_str_woodiness_analysis.R")
```

For the univartiate Metric ~ trait analysis, very little effect.
Phylogenetic effect is not well estimated.
Sample size is quite small.
This might suggest that traits can't be examined in isolation, either from eachother or from climate information.

144 - height
176 - woodiness
137 - life form
39 - Sla
80 - Seed mass
34 - leaf area

## 2. metric ~ phylogeny and trait ~ phylogeny
```{r}
#source("06_metric_phylogeny_analysis.R")
```

## 3. metric ~ multiple traits (*climate)
```{r}
#source("07_str_multitrait_analysis.R")
```


## 4. metric ~ humanfootprint
```{r}
#source("08a_str_get_hf_ratio.R")
#source("08b_str_humanfootprint_analysis.R")
```



## 5. plots
```{r}
# m_indiv_height <- readRDS("m_indiv/m_height_phy_parexp.rds")
# m_indiv_seed_mass <- readRDS("m_indiv/m_seed_mass_phylo_parexp.rds")
# m_indiv_sla <- readRDS("m_indiv/m_sla_phylo_parexp.rds")
# m_indiv_leaf_area <- readRDS("m_indiv/m_leaf_area.rds")
# m_indiv_woodiness <- readRDS("m_indiv/m_woodiness_phlyo_parexp.rds")
# m_indiv_lifeform <- readRDS("m_indiv/m_lifeform_phlyo_parexp.rds")
source("09_str_plots.R")
```
